cleanup struct
