+++ litestar/plugins/core/__init__.py
@@ -1,0 +1,3 @@
+from ._msgspec import MsgspecDIPlugin
+
+__all__ = ("MsgspecDIPlugin",)
